+++ BUILD.gn
@@ -1196,7 +1196,6 @@ config("grpc_config") {
         "include/grpcpp/server.h",
         "include/grpcpp/server_builder.h",
         "include/grpcpp/server_context.h",
-        "include/grpcpp/server_impl.h",
         "include/grpcpp/server_posix.h",
         "include/grpcpp/support/async_stream.h",
         "include/grpcpp/support/async_stream_impl.h",
