+++ gRPC-C++.podspec
@@ -164,7 +164,6 @@ Pod::Spec.new do |s|
                       'include/grpcpp/server.h',
                       'include/grpcpp/server_builder.h',
                       'include/grpcpp/server_context.h',
-                      'include/grpcpp/server_impl.h',
                       'include/grpcpp/server_posix.h',
                       'include/grpcpp/support/async_stream.h',
                       'include/grpcpp/support/async_stream_impl.h',
