+++ include/grpcpp/impl/codegen/async_generic_service.h
@@ -44,7 +44,7 @@ class GenericServerContext final : public ::grpc_impl::ServerContext {
   const std::string& host() const { return host_; }
 
  private:
-  friend class grpc_impl::Server;
+  friend class grpc::Server;
   friend class grpc::ServerInterface;
 
   void Clear() {
@@ -84,8 +84,8 @@ class AsyncGenericService final {
                    ::grpc::ServerCompletionQueue* notification_cq, void* tag);
 
  private:
-  friend class grpc_impl::Server;
-  grpc_impl::Server* server_;
+  friend class grpc::Server;
+  grpc::Server* server_;
 };
 
 #ifndef GRPC_CALLBACK_API_NONEXPERIMENTAL
@@ -105,7 +105,7 @@ class GenericCallbackServerContext final
   const std::string& host() const { return host_; }
 
  private:
-  friend class ::grpc_impl::Server;
+  friend class ::grpc::Server;
   friend class ::grpc::ServerInterface;
 
   void Clear() {
@@ -140,7 +140,7 @@ class CallbackGenericService {
   }
 
  private:
-  friend class ::grpc_impl::Server;
+  friend class grpc::Server;
 
   ::grpc_impl::internal::CallbackBidiHandler<ByteBuffer, ByteBuffer>*
   Handler() {
@@ -151,7 +151,7 @@ class CallbackGenericService {
         });
   }
 
-  grpc_impl::Server* server_{nullptr};
+  grpc::Server* server_{nullptr};
 };
 
 #ifndef GRPC_CALLBACK_API_NONEXPERIMENTAL
