+++ tools/doxygen/Doxyfile.c++
@@ -1024,7 +1024,6 @@ include/grpcpp/security/tls_credentials_options.h \
 include/grpcpp/server.h \
 include/grpcpp/server_builder.h \
 include/grpcpp/server_context.h \
-include/grpcpp/server_impl.h \
 include/grpcpp/server_posix.h \
 include/grpcpp/support/async_stream.h \
 include/grpcpp/support/async_stream_impl.h \
